NN-4892: Refactor
