+++ server/views/partials/hearingDetailsVariations/hearingTable.njk
@@ -2,16 +2,15 @@
 {% from "govuk/components/button/macro.njk" import govukButton %}
 
 {% macro hearingTable(hearingNumber, hearing, readOnly, reportNo, latestHearingId) %}
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-three-quarters">
-        {% set hearingId = hearing.id %}
-        <h2 class="govuk-heading-m" data-qa="hearing-index-{{hearingNumber}}">Hearing {{ hearingNumber }}</h2>
-        {% if not readOnly %}
-            {{ hearingSummary(hearing, { "data-qa": "details-summary-table-"+hearingNumber }, hearing.id === latestHearingId, adjudicationUrls.scheduleHearing.urls.edit(reportNo, hearingId), reportNo, latestHearingId) }}
-        {% else %}
-            {{ hearingSummary(hearing, { "data-qa": "details-summary-table-"+hearingNumber }, false) }}
-        {% endif %}
-        </div>
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-three-quarters">
+      {% set hearingId = hearing.id %}
+      <h2 class="govuk-heading-m" data-qa="hearing-index-{{hearingNumber}}">Hearing {{ hearingNumber }}</h2>
+      {% if not readOnly %}
+        {{ hearingSummary(hearing, { "data-qa": "details-summary-table-"+hearingNumber }, hearing.id === latestHearingId, adjudicationUrls.scheduleHearing.urls.edit(reportNo, hearingId), reportNo, latestHearingId) }}
+      {% else %}
+        {{ hearingSummary(hearing, { "data-qa": "details-summary-table-"+hearingNumber }, false) }}
+      {% endif %}
     </div>
+  </div>
 {% endmacro %}
-
