+++ server/views/partials/hearingDetailsVariations/hearingTableExpansions/hearingTableAdjourn.njk
@@ -1,0 +1,33 @@
+{% set newRows = [
+  {
+    key: {
+      text: 'Name of adjudicator'
+    },
+    value: {
+      text: hearingDetails.outcome.adjudicator | initialiseName
+    }
+  }, {
+    key: {
+      text: 'Next step'
+    },
+    value: {
+      text: 'Adjourn the hearing for another reason'
+    }
+  }, {
+    key: { text: 'Reason' },
+    value: {
+      html: '<p>'+hearingDetails.outcome.reason | convertHearingOutcomeAdjournReason +'</p><p>'+hearingDetails.outcome.details+'</p>'
+    }
+  },
+  {
+    key: { text: 'Plea' },
+    value: {
+      text: hearingDetails.outcome.plea | convertHearingOutcomePlea
+    }
+  }
+]
+ %}
+
+{% for individualRow in newRows %}
+  {% set rows = (rows.push(individualRow), rows) %}
+{% endfor %}
